Refactored the programs

+++ Basic_core_problems/PrimeFactors.php
@@ -8,7 +8,7 @@ $ans = 0;
 /**
  * Taking user input.
  */
-$num = (int)readline('Enter Number of which you want to find prime factors: ');
+$num = readline('Enter Number of which you want to find prime factors: ');
 for($i=2;$i<$num;$i++){
     while($num % $i == 0){
         echo $i . "\n";
